# specs/openapi.yaml
...
openapi: 3.1.0
info:
  title: dadeul API
  version: 0.1.0
  description: 다들 — 지역 기반 실시간 투표 위젯 서비스 API
servers:
  - url: https://api.dadeul.app
    description: production (예약)
  - url: http://localhost:3000
    description: local
...
paths:
...
  /v1/regions/resolve:
    post:
      operationId: resolveRegion
      tags: [region]
      summary: GPS 좌표 → 시군구 판별 (action-region-resolve)
      description: >
        좌표는 판별 즉시 폐기하며 저장하지 않는다. 시군구 코드가 저장 최소 단위.
        판별 결과는 계정의 캐시 지역으로 저장된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude: { type: number, minimum: -90, maximum: 90 }
                longitude: { type: number, minimum: -180, maximum: 180 }
      responses:
        "200":
          description: 판별 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "422":
          description: REGION_UNRESOLVED — 판별 불가 좌표(해외 등), 수동 선택 fallback 유도
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: bootstrapAccount가 발급한 계정 토큰
  responses:
...
    Unauthorized:
      description: 토큰 없음/무효
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: >
        RATE_LIMITED — 요청 상한 초과 (전역: 분당 계정/IP 기준, bootstrap·resolve는 별도 강화 상한).
        상한 값은 서버 설정(RATE_LIMIT_*_PER_MIN)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [token, message]
      properties:
        token:
          type: string
          description: 에러 토큰은 UX다 — 클라이언트는 토큰별 고유 안내·복구 동선으로 매핑
          enum:
            - TOPIC_NOT_FOUND
            - TOPIC_CLOSED
            - INVALID_OPTION
            - INSUFFICIENT_CREDIT
            - REGION_UNRESOLVED
            - REGION_INACTIVE
            - UNAUTHORIZED
            - RATE_LIMITED
            - BAD_REQUEST
        message:
          type: string
          description: 개발용 설명 — 유저 노출 문구는 클라이언트가 토큰으로 결정
...
    Region:
      type: object
      required: [code, name, fullName, active]
      properties:
        code: { type: string, description: 시군구 행정표준코드 (term-region-code) }
        name: { type: string, description: 단축 표시명 (강남구) }
        fullName: { type: string, description: 정식명 (서울특별시 강남구) }
        active: { type: boolean }
...
security:
  - bearerAuth: []